requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.6
Django==4.1.5
django-celery-beat==2.4.0
django-environ==0.9.0
django-redis==5.2.0
django-storages==1.13.2
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
et-xmlfile==1.1.0
filelock==3.9.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
numpy==1.24.1
openpyxl==3.0.10
packaging==23.0
pandas==1.5.3
pbr==5.11.1
platformdirs==2.6.2
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2022.7.1
redis==4.4.2
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
stevedore==4.1.1
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
virtualenv==20.17.1
virtualenv-clone==0.5.7
wcwidth==0.2.6
